Refactor and document ImageRecognition project
