--- ShopLens/ImageRecognition/Classificators/IImageClassifying.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace ImageRecognition.Classificators
-{
-    public interface IImageClassifying
-    {
-        Dictionary<string, float> ClassifyImage(byte[] image);
-    }
-}
