+++ ShopLens/ImageRecognition/Classificators/TensorFlowClassificator.cs
@@ -1,48 +1,57 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using ImageRecognition.Classificators.Helpers;
 using TensorFlow;
 
 namespace ImageRecognition.Classificators
 {
-    public class TensorFlowClassificator : IImageClassifying
+    /// <summary>
+    /// Image classificator based on the TensorFlow library.
+    /// </summary>
+    public class TensorFlowClassificator : IImageClassificator
     {
         private static string TensorFlowLabelsFilePath { get; } =
             Path.Combine("resources", "model", "labels.txt");
 
         private static string TensorFlowModelFilePath { get; } =
             Path.Combine("resources", "model", "model.pb");
-        
+
+        /// <summary>
+        /// Classifies a given tensor by using a model and label names.
+        /// </summary>
+        /// <param name="tensor">The tensor so classify. The tensor is usually created from an image.</param>
+        /// <param name="model"></param>
+        /// <param name="labels"></param>
+        /// <returns></returns>
         private Dictionary<string, float> ClassifyTensor(TFTensor tensor, byte[] model, IReadOnlyList<string> labels)
         {
+            // Recognition
             var graph = new TFGraph();
             graph.Import(model);
-
-            Dictionary<string, float> classificationResults;
+            TFTensor result;
             using (var session = new TFSession(graph))
             {
                 var runner = session.GetRunner();
                 runner.AddInput(graph["Placeholder"][0], tensor).Fetch(graph["loss"][0]);
                 var output = runner.Run();
-                var result = output[0];
-
-                // Results
-                var probabilities = ((float[][]) result.GetValue(jagged: true))[0];
-
-                classificationResults =
-                    probabilities
-                        .Select((x, i) => new KeyValuePair<string, float>(labels[i], probabilities[i]))
-                        .ToDictionary(x => x.Key, x => x.Value);
+                result = output[0];
             }
 
+            // Results
+            var probabilities = ((float[][]) result.GetValue(jagged: true))[0];
+
+            var classificationResults = probabilities
+                .Select((x, i) => new KeyValuePair<string, float>(labels[i], probabilities[i]))
+                .ToDictionary(x => x.Key, x => x.Value);
             return classificationResults;
         }
 
+        /// <inheritdoc cref="IImageClassificator.ClassifyImage"/>
         public Dictionary<string, float> ClassifyImage(byte[] image)
         {
             var tensor = TfImageUtil.CreateTensorFromBytes(image);
-            
-            // Use default settings for now
+
             var model = File.ReadAllBytes(TensorFlowModelFilePath);
             var labels = File.ReadAllLines(TensorFlowLabelsFilePath);
